requirements.txt
pycups
json
my-magento